refactor: Remove "Buscar por universidad" link from layout component
The "Buscar por universidad" link has been removed from the layout component HTML file. This change is in line with the recent user commits that updated the map component to display universities as markers. The link is no longer needed and has been removed to improve the overall user experience.